+++ ext/ext.go
@@ -1,0 +1,2 @@
+// Package ext contains some extra code which does not fit into the categorization of the stdlib.
+package ext